+++ pakyow-core/lib/pakyow/behavior/verification.rb
@@ -1,17 +1,14 @@
 # frozen_string_literal: true
 
 require "pakyow/support/class_state"
-require "pakyow/support/deep_dup"
 require "pakyow/support/extension"
 
-require "pakyow/errors"
 require "pakyow/verifier"
 
 module Pakyow
   module Behavior
     module Verification
       extend Support::Extension
-      using Support::DeepDup
 
       def verify(values = nil, &block)
         unless values
@@ -22,14 +19,7 @@ module Pakyow
           end
         end
 
-        original_values = values.deep_dup
-        result = Pakyow::Verifier.new(&block).call(values, context: self)
-
-        unless result.verified?
-          error = InvalidData.new_with_message(:verification)
-          error.context = { object: original_values, result: result }
-          raise error
-        end
+        Pakyow::Verifier.new(&block).call!(values, context: self)
       end
 
       apply_extension do